requirements.txt
asgiref==3.2.3
Django==3.0.1
django-ckeditor==5.8.0
django-js-asset==1.2.2
mysqlclient==1.4.6
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.0
